+++ src/components/CreateFlow.js
@@ -2,6 +2,7 @@
 //https://docs.superfluid.finance/superfluid/developers/constant-flow-agreement-cfa/money-streaming-1
 
 import React, { useState } from "react";
+import { Framework } from "@superfluid-finance/sdk-core";
 import {
   Button,
   Form,
@@ -15,7 +16,17 @@ import { ethers } from "ethers";
 //where the Superfluid logic takes place
 async function createNewFlow(recipient, flowRate) {
 
-  signer, sf = await getSfFramework(window);
+  const provider = new ethers.providers.Web3Provider(window.ethereum);
+  const signer = provider.getSigner();
+
+  console.log(signer._address);
+
+  const chainId = await window.ethereum.request({ method: "eth_chainId" });
+
+  const sf = await Framework.create({
+      chainId: Number(chainId),
+      provider: provider
+  });
 
   const DAIxContract = await sf.loadSuperToken("fDAIx");
   const DAIx = DAIxContract.address;
@@ -83,9 +94,13 @@ export const CreateFlow = () => {
   };
 
   const handleFlowRateChange = (e) => {
-    setFlowRate(() => ([e.target.name] = e.target.value));
-    let newFlowRateDisplay = calculateFlowRate(e.target.value);
-    setFlowRateDisplay(newFlowRateDisplay.toString());
+    try {
+      setFlowRate(() => ([e.target.name] = e.target.value));
+      let newFlowRateDisplay = calculateFlowRate(e.target.value);
+      setFlowRateDisplay(newFlowRateDisplay.toString());
+    } catch {
+      console.error("Flowrate invalid.")
+    }
   };
 
   return (
